refactor: introduce a single generateRandomId
